fixed uninitialized member variables in Cexif constructor.
cppcheck reported as:
Message: Member variable 'Cexif::m_exifinfo' is not initialized in the constructor.
Message: Member variable 'Cexif::m_szLastError' is not initialized in the constructor.
Message: Member variable 'Cexif::ExifImageWidth' is not initialized in the constructor.
Message: Member variable 'Cexif::MotorolaOrder' is not initialized in the constructor.
Message: Member variable 'Cexif::Sections' is not initialized in the constructor.
Message: Member variable 'Cexif::SectionsRead' is not initialized in the constructor.
Message: Member variable 'Cexif::freeinfo' is not initialized in the constructor.